+++ src/main/java/DGU_AI_LAB/admin_be/domain/requests/controller/AdminRequestController.java
@@ -7,7 +7,8 @@ import DGU_AI_LAB.admin_be.domain.requests.dto.response.ChangeRequestResponseDTO
 import DGU_AI_LAB.admin_be.domain.requests.dto.response.ContainerInfoDTO;
 import DGU_AI_LAB.admin_be.domain.requests.dto.response.ResourceUsageDTO;
 import DGU_AI_LAB.admin_be.domain.requests.dto.response.SaveRequestResponseDTO;
-import DGU_AI_LAB.admin_be.domain.requests.service.AdminRequestService;
+import DGU_AI_LAB.admin_be.domain.requests.service.AdminRequestCommandService;
+import DGU_AI_LAB.admin_be.domain.requests.service.AdminRequestQueryService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -21,25 +22,26 @@ import java.util.List;
 @RequestMapping("/api/admin/requests")
 public class AdminRequestController {
 
-    private final AdminRequestService adminRequestService;
+    private final AdminRequestCommandService adminRequestCommandService;
+    private final AdminRequestQueryService adminRequestQueryService;
 
     @PatchMapping("/change/approve")
     public ResponseEntity<Void> approveModification(
             @AuthenticationPrincipal(expression = "userId") Long adminId,
             @RequestBody @Valid ApproveModificationDTO dto
     ) {
-        adminRequestService.approveModification(adminId, dto);
+        adminRequestCommandService.approveModification(adminId, dto);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/usage")
     public ResponseEntity<List<ResourceUsageDTO>> getAllResourceUsage() {
-        return ResponseEntity.ok(adminRequestService.getAllFulfilledResourceUsage());
+        return ResponseEntity.ok(adminRequestQueryService.getAllFulfilledResourceUsage());
     }
 
     @GetMapping("/containers")
     public ResponseEntity<List<ContainerInfoDTO>> getAllActiveContainers() {
-        return ResponseEntity.ok(adminRequestService.getAllActiveContainers());
+        return ResponseEntity.ok(adminRequestQueryService.getAllActiveContainers());
     }
 
     /**
@@ -48,7 +50,7 @@ public class AdminRequestController {
      */
     @GetMapping("/new")
     public ResponseEntity<List<SaveRequestResponseDTO>> getNewRequests() {
-        List<SaveRequestResponseDTO> requests = adminRequestService.getNewRequests();
+        List<SaveRequestResponseDTO> requests = adminRequestQueryService.getNewRequests();
         return ResponseEntity.ok(requests);
     }
 
@@ -58,17 +60,17 @@ public class AdminRequestController {
      */
     @GetMapping("/change")
     public ResponseEntity<List<ChangeRequestResponseDTO>> getChangeRequests() {
-        List<ChangeRequestResponseDTO> changeRequests = adminRequestService.getChangeRequests();
+        List<ChangeRequestResponseDTO> changeRequests = adminRequestQueryService.getChangeRequests();
         return ResponseEntity.ok(changeRequests);
     }
 
     @PatchMapping("/approval")
     public ResponseEntity<SaveRequestResponseDTO> approve(@RequestBody @Valid ApproveRequestDTO dto) {
-        return ResponseEntity.ok(adminRequestService.approveRequest(dto));
+        return ResponseEntity.ok(adminRequestCommandService.approveRequest(dto));
     }
 
     @PatchMapping("/reject")
     public ResponseEntity<SaveRequestResponseDTO> reject(@RequestBody @Valid RejectRequestDTO dto) {
-        return ResponseEntity.ok(adminRequestService.rejectRequest(dto));
+        return ResponseEntity.ok(adminRequestCommandService.rejectRequest(dto));
     }
 }
